Changed signatures for MacFUSE-specific set/getxattr functions. Last argument is now a jlong instead of jint so it can represent all values of uint32_t correctly. Also updated compiled library.

+++ libjfuse/src/fuse26_module.cpp
@@ -616,7 +616,7 @@ int jfuse_setxattr(const char *path, const char *name, const char *value,
 
     JFUSE_FS_PROVIDER_MID_OK(OPS_SETXATTR_BSD_NAME, OPS_SETXATTR_BSD_SIGNATURE) {
 
-        JFUSE_FS_PROVIDER_CALL(JAVA_ARG(1), JAVA_ARG(2), JAVA_ARG(3), flags, position);
+        JFUSE_FS_PROVIDER_CALL(JAVA_ARG(1), JAVA_ARG(2), JAVA_ARG(3), (jint)flags, (jlong)position);
 
         JFUSE_SET_RETVAL();
     }
@@ -648,7 +648,7 @@ int jfuse_setxattr(const char *path, const char *name, const char *value,
 
     JFUSE_FS_PROVIDER_MID_OK(OPS_SETXATTR_NAME, OPS_SETXATTR_SIGNATURE) {
 
-        JFUSE_FS_PROVIDER_CALL(JAVA_ARG(1), JAVA_ARG(2), JAVA_ARG(3), flags);
+        JFUSE_FS_PROVIDER_CALL(JAVA_ARG(1), JAVA_ARG(2), JAVA_ARG(3), (jint)flags);
 
         JFUSE_SET_RETVAL();
     }
@@ -683,7 +683,7 @@ int jfuse_getxattr(const char *path, const char *name, char *value,
 
     JFUSE_FS_PROVIDER_MID_OK(OPS_GETXATTR_BSD_NAME, OPS_GETXATTR_BSD_SIGNATURE) {
 
-        JFUSE_FS_PROVIDER_CALL(JAVA_ARG(1), JAVA_ARG(2), JAVA_ARG(3), position);
+        JFUSE_FS_PROVIDER_CALL(JAVA_ARG(1), JAVA_ARG(2), JAVA_ARG(3), (jlong)position);
 
         JFUSE_SET_RETVAL();
     }
